Replace nodePalette util wit registry lookup
AddCountButton now takes a color name (e.g., 'kiwi') corresponding to
one of the named colors in Network-Canvas-UI.
