+++ src/selectors/protocol.js
@@ -2,7 +2,26 @@
 
 import { createDeepEqualSelector } from './utils';
 
+export const protocolRegistry = createDeepEqualSelector(
+  state => state.protocol.variableRegistry,
+  registry => registry,
+);
+
+export const protocolForms = createDeepEqualSelector(
+  state => state.protocol.forms,
+  forms => forms,
+);
+
 export const getExternalData = createDeepEqualSelector(
   state => state.protocol.externalData,
   protocolData => protocolData,
+);
+
+export const getNodeColor = createDeepEqualSelector(
+  protocolRegistry,
+  (state, props) => props.nodeType,
+  (variableRegistry, nodeType) => {
+    const nodeInfo = variableRegistry.node;
+    return nodeInfo && nodeInfo[nodeType] && nodeInfo[nodeType].color;
+  },
 );
