+++ src/selectors/rehydrate.js
@@ -1,14 +1,7 @@
 /* eslint-disable import/prefer-default-export */
 
-import { createSelector, createSelectorCreator, defaultMemoize } from 'reselect';
-import { isEqual } from 'lodash';
-
-// create a "selector creator" that uses lodash.isEqual instead of ===
-
-const createDeepEqualSelector = createSelectorCreator(
-  defaultMemoize,
-  isEqual,
-);
+import { createSelector } from 'reselect';
+import { protocolRegistry, protocolForms } from './protocol';
 
 // Prop selectors
 
@@ -19,15 +12,6 @@ const propForm = (_, { entity, type }) => ({ entity, type });
 
 // MemoedSelectors
 
-export const protocolRegistry = createDeepEqualSelector(
-  state => state.protocol.variableRegistry,
-  registry => registry,
-);
-
-export const protocolForms = createDeepEqualSelector(
-  state => state.protocol.forms,
-  forms => forms,
-);
 
 const rehydrateField = ({ registry, entity, type, field }) => {
   if (!field.variable) { return field; }
