+++ src/styles/containers/_interfaces.scss
@@ -1,6 +1,4 @@
-/**
- * Shared mixins forinterfaces.
- */
+// Shared mixins for interfaces.
 
 @mixin interface-prompt {
   flex: 0 0 12vw;
